# openapi_output.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smith 2.0",
        "description": "Agente de Vendas Inteligente com IA",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/analytics/dashboard": {
            "get": {
                "tags": [
                    "Analytics",
                    "Analytics"
                ],
                "summary": "Get Dashboard Metrics",
                "description": "Retorna m\u00c3\u00a9tricas do dashboard em tempo real\n\nArgs:\n    periodo_dias: N\u00c3\u00bamero de dias para an\u00c3\u00a1lise (1-365)\n\nReturns:\n    M\u00c3\u00a9tricas completas do dashboard:\n    - Resumo geral (total leads, pipeline value, crescimento)\n    - Funil de convers\u00c3\u00a3o por etapa\n    - Distribui\u00c3\u00a7\u00c3\u00a3o por temperatura\n    - Tempo m\u00c3\u00a9dio em cada est\u00c3\u00a1gio\n    - Motivos de perda\n    - Timeline de evolu\u00c3\u00a7\u00c3\u00a3o\n    - Taxas de convers\u00c3\u00a3o",
                "operationId": "get_dashboard_metrics_api_analytics_dashboard_get",
                "parameters": [
                    {
                        "name": "periodo_dias",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1,
                            "description": "Per\u00c3\u00adodo em dias para an\u00c3\u00a1lise",
                            "default": 30,
                            "title": "Periodo Dias"
                        },
                        "description": "Per\u00c3\u00adodo em dias para an\u00c3\u00a1lise"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Dashboard Metrics Api Analytics Dashboard Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/funil": {
            "get": {
                "tags": [
                    "Analytics",
                    "Analytics"
                ],
                "summary": "Get Funil Conversao",
                "description": "Retorna dados do funil de convers\u00c3\u00a3o\n\nArgs:\n    periodo_dias: Per\u00c3\u00adodo em dias\n\nReturns:\n    Dados do funil com contagem e percentual por etapa",
                "operationId": "get_funil_conversao_api_analytics_funil_get",
                "parameters": [
                    {
                        "name": "periodo_dias",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1,
                            "default": 30,
                            "title": "Periodo Dias"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Funil Conversao Api Analytics Funil Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/analytics/timeline": {
            "get": {
                "tags": [
                    "Analytics",
                    "Analytics"
                ],
                "summary": "Get Timeline",
                "description": "Retorna timeline de evolu\u00c3\u00a7\u00c3\u00a3o de leads\n\nArgs:\n    periodo_dias: Per\u00c3\u00adodo em dias (7-90)\n\nReturns:\n    S\u00c3\u00a9rie temporal com novos leads, qualificados e perdidos por dia",
                "operationId": "get_timeline_api_analytics_timeline_get",
                "parameters": [
                    {
                        "name": "periodo_dias",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 90,
                            "minimum": 7,
                            "default": 30,
                            "title": "Periodo Dias"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Timeline Api Analytics Timeline Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/analytics/motivos-perda": {
            "get": {
                "tags": [
                    "Analytics",
                    "Analytics"
                ],
                "summary": "Get Motivos Perda",
                "description": "Retorna principais motivos de perda de leads\n\nArgs:\n    periodo_dias: Per\u00c3\u00adodo em dias\n\nReturns:\n    Lista de motivos com contagem e percentual",
                "operationId": "get_motivos_perda_api_analytics_motivos_perda_get",
                "parameters": [
                    {
                        "name": "periodo_dias",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1,
                            "default": 30,
                            "title": "Periodo Dias"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Motivos Perda Api Analytics Motivos Perda Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}